requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
bcrypt==4.0.1
xlrd==2.0.1